split out AddLayerDialog class into its own file
